# src/main/java/com/translation/translationservicemgmt/dto/TranslationResponseDto.java
package com.translation.translationservicemgmt.dto;

import java.util.List;

public class TranslationResponseDto {

    private Long id;
    private String key;
    private String locale;
    private String content;
    private List<String> tags;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public List<String> getTags() {
        return tags;
    }

    public void setTags(List<String> tags) {
        this.tags = tags;
    }

    public String getContent() {
        return content;
    }

    public void setContent(String content) {
        this.content = content;
    }

    public String getLocale() {
        return locale;
    }

    public void setLocale(String locale) {
        this.locale = locale;
    }

    public String getKey() {
        return key;
    }

    public void setKey(String key) {
        this.key = key;
    }
}